refactor: reduce swap/bridge tx mock data
